Autoconverted sources with fixes ydanila/sharpen_imazen@f26398219e8ab7215985a193a8e1aeefc660c7d0

+++ Com.Drew/Com/Drew/Metadata/Directory.cs
@@ -616,7 +616,7 @@ namespace Com.Drew.Metadata
 			double? value = GetDoubleObject(tagType);
 			if (value != null)
 			{
-				return value;
+				return (double)value;
 			}
 			object o = GetObject(tagType);
 			if (o == null)
